[refactor] move file for navigation
