+++ app/src/main/java/com/mashup/ui/moremenu/MoreMenuViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.moremenu
+package com.mashup.ui.moremenu
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
